requirements.txt
wheel==0.31.0
six==1.11.0
setuptools==39.0.1
pytz==2018.3
python-mimeparse==1.6.0
python-dateutil==2.7.2
pip==9.0.3
djangorestframework==3.7.7
Django==1.11.11
django-tastypie==0.14.0
zict==0.1.2
xlwt==1.2.0
xlwings==0.10.4
XlsxWriter==0.9.6
xlrd==1.0.0
wrapt==1.10.10
win-unicode-console==0.5
widgetsnbextension==2.0.0
Werkzeug==0.12.2
wcwidth==0.1.7
virtualenv==15.2.0
uritemplate==3.0.0
unicodecsv==0.14.1
tslearn==0.1.8
traitlets==4.3.2
tornado==4.5.1
toolz==0.8.2
Theano==0.9.0
testpath==0.3
tblib==1.3.2
tables==3.2.2
sympy==1.0
subprocess32==3.2.7
statsmodels==0.8.0
SQLAlchemy==1.1.9
spyder==3.1.4
sortedcontainers==1.5.7
sortedcollections==0.5.3
snowballstemmer==1.2.1
smart-open==1.5.3
singledispatch==3.4.0.3
simplegeneric==0.8.1
seaborn==0.7.1
scipy==0.19.0
scikit-neuralnetwork==0.7
scikit-multilearn==0.0.5
scikit-learn==0.18.1
scikit-image==0.13.0
scandir==1.5
rope==0.9.4
requests==1.1.0
QtPy==1.2.1
qtconsole==4.3.0
QtAwesome==0.4.4
pyzmq==16.0.2
PyYAML==3.12
pywin32==220
PyWavelets==0.5.2
python-mnist==0.3
python-firebase==1.2
pytest==3.0.7
pyparsing==2.1.4
pyOpenSSL==17.0.0
pyodbc==4.0.16
pylint==1.6.4
Pygments==2.2.0
pyflakes==1.5.0
pycurl==7.43.0
pycrypto==2.6.1
pycparser==2.17
pycosat==0.6.2
py==1.4.33
psutil==5.2.2
prompt-toolkit==1.0.14
ply==3.10
plotly==2.0.16
Pillow==4.1.1
pickleshare==0.7.4
pep8==1.7.0
patsy==0.4.1
pathlib2==2.2.1
path.py==10.3.1
partd==0.3.8
pandocfilters==1.4.1
pandas==0.20.1
packaging==16.8
openpyxl==2.4.7
olefile==0.44
odo==0.5.0
numpydoc==0.6.0
numpy==1.12.1
numexpr==2.6.2
numba==0.33.0+0.ge79330a.dirty
notebook==5.0.0
nose==1.3.7
nltk==3.2.3
networkx==1.11
nbformat==4.3.0
nbconvert==5.1.1
navigator-updater==0.1.0
multipledispatch==0.4.9
msgpack-python==0.4.8
mpmath==0.19
mistune==0.7.4
menuinst==1.4.7
matplotlib==2.0.2
MarkupSafe==0.23
lxml==3.7.3
locket==0.2.0
llvmlite==0.18.0
lazy-object-proxy==1.2.2
Lasagne==0.1
jupyter==1.0.0
jupyter-core==4.3.0
jupyter-console==5.1.0
jupyter-client==5.0.1
jsonschema==2.6.0
Jinja2==2.9.6
jedi==0.10.2
jdcal==1.3
itypes==1.1.0
itsdangerous==0.24
isort==4.2.5
ipywidgets==6.0.0
ipython==5.3.0
ipython-genutils==0.2.0
ipykernel==4.6.1
ipaddress==1.0.18
imagesize==0.7.1
idna==2.5
html5lib==0.999
HeapDict==1.0.0
h5py==2.7.0
grin==1.2.1
greenlet==0.4.12
gevent==1.2.1
gensim==2.3.0
futures==3.1.1
future==0.16.0
functools32==3.2.3.post2
funcsigs==1.0.2
Flask==0.12.2
Flask-Cors==3.0.2
fastcache==1.0.2
et-xmlfile==1.0.1
enum34==1.1.6
entrypoints==0.2.2
docutils==0.13.1
distributed==1.16.3
decorator==4.0.11
datashape==0.5.4
dask==0.14.3
cytoolz==0.8.2
Cython==0.25.2
cycler==0.10.0
cryptography==1.8.1
coreschema==0.0.4
coreapi==2.3.3
contextlib2==0.5.5
configparser==3.5.0
conda==4.3.30
comtypes==1.1.2
colorama==0.3.9
clyent==1.2.2
cloudpickle==0.2.2
click==6.7
chardet==3.0.3
cffi==1.10.0
cdecimal==2.3
bz2file==0.98
Bottleneck==1.2.1
boto==2.46.1
bokeh==0.12.5
bleach==1.5.0
blaze==0.10.1
bitstring==3.1.5
bitarray==0.8.1
beautifulsoup4==4.6.0
backports.ssl-match-hostname==3.4.0.2
backports.shutil-get-terminal-size==1.0.0
backports-abc==0.5
Babel==2.4.0
astropy==1.3.2
astroid==1.4.9
asn1crypto==0.22.0
anaconda-project==0.6.0
anaconda-navigator==1.6.2
anaconda-client==1.6.3
alabaster==0.7.10
sphinx==1.5.6
